requirements.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
beautifulsoup4==4.6.0
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
dill==0.2.7.1
enum34==1.1.6
future==0.16.0
futures==3.1.1
google-auth==1.1.1
google-cloud-core==0.27.1
google-cloud-vision==0.27.0
google-gax==0.15.15
googleapis-common-protos==1.5.3
grpcio==1.4.0
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
olefile==0.44
Pillow==4.3.0
ply==3.8
protobuf==3.4.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
pylint==1.7.4
requests==2.18.4
rsa==3.4.2
singledispatch==3.4.0.3
six==1.11.0
urllib3==1.22
wrapt==1.10.11
